+++ ChangeLog.md
@@ -1,3 +1,10 @@
+02-Jan-2015  Alexandre Trilla  <alex@atrilla.net>
+
+    * src/breadth_first_search.m: created.
+
+    * test/prettyplot.m: former ut file.
+
+
 26-Dec-2014  Alexandre Trilla  <alex@atrilla.net>
 
     * src/ut_prettyplot.m: created.